requirements.txt
flask
pyserial
requests